[skip ci] WIP on rational.
