+++ tests/CMakeLists.txt
@@ -61,8 +61,6 @@ ADD_PIRANHA_TESTCASE(math)
 ADD_PIRANHA_TESTCASE(memory)
 ADD_PIRANHA_TESTCASE(monomial_01)
 ADD_PIRANHA_TESTCASE(monomial_02)
-ADD_PIRANHA_TESTCASE(mp_rational_01)
-ADD_PIRANHA_TESTCASE(mp_rational_02)
 ADD_PIRANHA_TESTCASE(parallel_vector_transform)
 ADD_PIRANHA_TESTCASE(poisson_series_01)
 ADD_PIRANHA_TESTCASE(poisson_series_02)
@@ -81,6 +79,8 @@ ADD_PIRANHA_TESTCASE(power_series_01)
 ADD_PIRANHA_TESTCASE(power_series_02)
 ADD_PIRANHA_TESTCASE(print_coefficient)
 ADD_PIRANHA_TESTCASE(print_tex_coefficient)
+ADD_PIRANHA_TESTCASE(rational_01)
+ADD_PIRANHA_TESTCASE(rational_02)
 ADD_PIRANHA_TESTCASE(real_01)
 ADD_PIRANHA_TESTCASE(real_02)
 ADD_PIRANHA_TESTCASE(real_trigonometric_kronecker_monomial_01)
