+++ tests/rational_02.cpp
